feat(workflow): 新增财务统计功能并集成到多个业务流程
- 在 DcCustomerRefundMapper 中新增 insertFinancialStatistics 方法用于插入财务统计记录
- 在 DcCustomerRefundServiceImpl 中实现客户退费时自动生成财务统计记录
- 在 DcCustomerTransferController 中实现客户实收时生成财务统计记录
- 在 DcExpenseReimbursementServiceImpl 中实现员工报销完成时生成财务统计记录
- 更新财务统计表字段 balance 从 Long 改为 BigDecimal 类型
- 为财务统计对象添加金额、财务类型、来源类型等字段的验证注解
- 在工作流模块中添加 ruoyi-admin 依赖以支持财务统计服务调用

+++ ruoyi-admin/src/main/java/org/dromara/financial/domain/DcFinancialStatistics.java
@@ -1,15 +1,14 @@
 package org.dromara.financial.domain;
 
-import org.dromara.common.mybatis.core.domain.BaseEntity;
-import com.baomidou.mybatisplus.annotation.*;
+import com.baomidou.mybatisplus.annotation.TableId;
+import com.baomidou.mybatisplus.annotation.TableName;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
-import org.dromara.common.translation.annotation.Translation;
-import java.util.Date;
-import com.fasterxml.jackson.annotation.JsonFormat;
-import org.dromara.common.translation.constant.TransConstant;
+import org.dromara.common.mybatis.core.domain.BaseEntity;
 
 import java.io.Serial;
+import java.math.BigDecimal;
+import java.util.Date;
 
 /**
  * 财务统计对象 dc_financial_statistics
@@ -34,7 +33,7 @@ public class DcFinancialStatistics extends BaseEntity {
     /**
      * 金额
      */
-    private Long balance;
+    private BigDecimal balance;
 
     /**
      * 财务类型
